backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dotenv==1.0.0
django-cors-headers==4.1.0
dj-rest-auth==3.0.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
django-filter==23.2
drf-yasg==1.21.7
celery==5.3.1
redis==4.6.0
django-celery-beat==2.5.0